use "size" rather than "scale"
This is for better fine-tuning of the size of the separator. Also, the
previous workaround we had (the "line_height" hack) doesn't work in
sway, since it adjusts the total height of the bar, and eventually
exceeds the available window space, causing sway's renderer to break.

This current approach seems to work more reliably, and works across Xorg
as well as sway.